requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.8

# improved REPL
ipdb==0.10.3

flake8==3.5.0
coverage==4.4.1
